% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_res.R
\name{gg_res}
\alias{gg_res}
\title{get a residual plot}
\usage{
gg_res(df, .time, .res, .ct = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{.time}{time variable}

\item{.res}{residuals to plot}

\item{.ct}{central tendency}
}
\description{
get a residual plot
}
\details{
This function enables you to plot different types of residuals versus a time variable of choice
}
\examples{
\dontrun{gg_res(data, TAD, CWRES)}
}
